refactor(cloud): extract workflow path parsing into workflow-paths.ts
Move the pure, side-effect-free workflow-path parsers out of workflows.ts
(1129 -> 893 lines) into a focused workflow-paths.ts module: the YAML path
parser, the TypeScript literal scanner, the parseWorkflowPaths dispatcher, and
inferWorkflowFileType / shouldSyncCodeByDefault.

Validation (validateYamlWorkflow / validateTypeScriptWorkflow, which shells out),
the runtime PATH_NAME_RE guard, GitHub-remote helpers, and the cloud
execution/API/S3 code stay in workflows.ts, which re-exports the three parser
entry points so the package barrel and tests still import them from
'./workflows.js'. Pure move: no behavior change. Verified with build, full test
suite (959 pass), and a parser smoke test.
