+++ .agentworkforce/trajectories/active/traj_q768ixcuyzh7/trajectory.json
@@ -31,6 +31,18 @@
             "reasoning": "normalize.ts readers have divergent semantics (readString coerces number/bigint->string, readRecord clones) so consolidating would be a behavior change, not a pure refactor"
           },
           "significance": "high"
+        },
+        {
+          "ts": 1782055254612,
+          "type": "decision",
+          "content": "Extracted pure workflow-path parsing (YAML + TS literal scanner) from cloud/workflows.ts into workflow-paths.ts: Extracted pure workflow-path parsing (YAML + TS literal scanner) from cloud/workflows.ts into workflow-paths.ts",
+          "raw": {
+            "question": "Extracted pure workflow-path parsing (YAML + TS literal scanner) from cloud/workflows.ts into workflow-paths.ts",
+            "chosen": "Extracted pure workflow-path parsing (YAML + TS literal scanner) from cloud/workflows.ts into workflow-paths.ts",
+            "alternatives": [],
+            "reasoning": "workflows.ts mixed pure text parsing with auth/S3/API IO; isolating the side-effect-free parsers makes them testable and shrinks the IO module. Validation (incl child_process) and runtime-only regex kept behind."
+          },
+          "significance": "high"
         }
       ]
     }
